migrations: add inline_literal where applicable
This is a bit unfortunate, but this is how migrations are supposed to
be written: unadorned literals will emit a `UPDATE .. WHERE x=%()s`
sql, which of course doesn't have the actual values rendered.

This means they work at runtime (when applying them via alembic to a
live db), but rendering to sql (with --sql) renders unusable output,
because the output doesn't have the actual values.

Instead, we must wrap the literals in a `inline_literal`, and then
it'll be inlined into the resulting sql. We only need to do this in
migrations, don't worry about regular sqlalchemy usage. But in
migrations, this is just a hoop we have to jump through.