migrations: config tables: insert a json 'null', not a sql NULL
Saying `value=None` works when running online, because it apparently
goes through JSONString.process_bind_param and it dumps the value,
but not offline (`--sql`), because then it renders just a NULL.

And that column is `NOT NULL`.

If we explicitly say that it's a "null" literal, then it is a json
null indeed... but online, JSONString must still be able to handle
it. So let's just put in a specialcase for this.

I hate it, but that's what we have to do, to support both online
and offline.

Anyway, this ties back to how we use a text column instead of a json
column, so we have to do things by hand :/ But that's not really
changeable before 7.0