RD-4306 Allow migrations to be rendered to SQL (#3538)
* Remove the pointless "COMMIT" calls

I'm unable to see why are they even there.
But I want migrations to run in a single transaction, as designed

* migrations: add inline_literal where applicable

This is a bit unfortunate, but this is how migrations are supposed to
be written: unadorned literals will emit a `UPDATE .. WHERE x=%()s`
sql, which of course doesn't have the actual values rendered.

This means they work at runtime (when applying them via alembic to a
live db), but rendering to sql (with --sql) renders unusable output,
because the output doesn't have the actual values.

Instead, we must wrap the literals in a `inline_literal`, and then
it'll be inlined into the resulting sql. We only need to do this in
migrations, don't worry about regular sqlalchemy usage. But in
migrations, this is just a hoop we have to jump through.

* migrations: config tables: insert a json 'null', not a sql NULL

Saying `value=None` works when running online, because it apparently
goes through JSONString.process_bind_param and it dumps the value,
but not offline (`--sql`), because then it renders just a NULL.

And that column is `NOT NULL`.

If we explicitly say that it's a "null" literal, then it is a json
null indeed... but online, JSONString must still be able to handle
it. So let's just put in a specialcase for this.

I hate it, but that's what we have to do, to support both online
and offline.

Anyway, this ties back to how we use a text column instead of a json
column, so we have to do things by hand :/ But that's not really
changeable before 7.0

* Remove trailing whitespace

My editor does this automatically. No reason. But since it already did
it, might as well commit it. It's better anyway.