RD-4306 Migrations: do `alter type` outside transactions
Look, the `COMMIT` calls, that I got rid of in #3538, were required
after all!

Turns out, `alter type` must not run in a transaction:
```
sqlalchemy.exc.InternalError: (psycopg2.errors.ActiveSqlTransaction) ALTER TYPE ... ADD cannot run inside a transaction block
```

Unfortunate. But at least, instead of a bare `COMMIT`, we can use this
feature, that alembic added _specifically_ for this case.

It at least suggests that this is the command that must run outside
transactions, as opposed to a seemingly-unrelated bare `COMMIT`.

ref https://alembic.sqlalchemy.org/en/latest/api/runtime.html#alembic.runtime.migration.MigrationContext.autocommit_block