rewrite test_workflows to not use create-dep-env
We don't really care about creating the deployments here, just so that
they have the workflows we need them to have.

So skip all the clutter, and let's just create deps with the workflows
directly.
And then keep using the list and filter calls as usual.

This speeds the module up by some 80%