RD-4328 Allow filtering graphs by just execution-id (#3555)
* RD-4328 Allow filtering graphs by just execution-id

No need to require the name. We can allow to just list all graphs
for the given execution.

* tests: drop duplicated lines

well those are already in the mixin!

* tasks graph: use the CreatedAtMixin

the mixin has a python-side default, so let's use it. No need to
declare the column separately here.

sql-side, the column is the same:
```
[root@4acdc34e33d9 migrations]# /opt/manager/env/bin/alembic revision --autogenerate
[root@4acdc34e33d9 migrations]#
```
no migrations generated