RD-4292 Add more ways to filter listing operations (#3556)
* RD-4292 Add more ways to filter listing operations

Those are going to be useful for the user, to be able to list the
operations based on... those things.

* remove now-invalid test case

this is now prevented by the restclient itself

* tests config: also set default_page_size

I dont know why it only fails sometimes. But it should probably be set.

* when delattr'ing a setting, revert to the default

instead of to not-set. This makes it so that delattring reverts the
state to as it was before setting it.