RD-4410 Permissions api: handle duplicated permissions (#3560)
* Add base tests for the permissions endpoint

I'm going to be changing it, so it had better have baseline tests.

Also, add that one backref, for convenience.

(hey, 100% coverage of that module)

* RD-4410 Disallow adding the same permission multiple times

Unfortunately, doing this on the db side, would cause a nasty
back-compat issue, because people might have already introduced
duplicated permissions. Oh well, let's do it on the python side.

* RD-4410 Allow deleting duplicated permissions

If someone's gotten themselves into this silly situation, we'd better
give them a way out!