RD-3903 Keep custom IDDs after dep-update (#3561)
Just look at the idd's dependency_creator, and if it's not coming
from a function, it means it's created by the user explicitly.
And in that case, we keep it.

I do realize that this isn't bulletproof, ie. someone might've created
an idd with "get_capability" in the name, but at that point we're so
deep into direct inner workings of the manager, that I don't even.

Like, the user knows what they're doing at that point, because tbh
that user is likely me anyway. So, suck it, me: better stay in line
and name your dependency creators in a smart way!