RD-4433 Config: refactor getting async db_url (#3569)
- there's 3 similar functions, so let's refactor them to all take the
  settings from the same base two sources, _get_sql_params and
  _render_db_url - I think that way is nicer
- the async ones don't seem to like connect_timeout
- the sqlalchemy-async one wants 'ssl' rather than 'sslmode'. I don't
  know why.

This makes this inte-test pass: integration_tests/tests/async_tests/test_audit_log.py
I'm not sure why haven't we seen that failure on jenkins. I don't think
we run that group there?

Anyway, this only fixes the "no-db-ssl" part. "db-with-ssl" is still
borked, because 'sslcert' and 'sslkey' aren't valid options for
sqlalchemy+asyncpg. It seems to me that it is just 100% borked and
needs to be fixed sqlalchemy-side.