RD-4455 Rewrite the flaky tests under ExecutionQueueingTests (#3577)
It's not guaranteed that the two executions would both be dequeued-
dequeueing works on a best-effort basis, wrt. amount of dequeued.

Although I don't know why it started to fail much more only recently.
Has to be something about timing.

Anyway, now the tests are more unit-y, in that they more directly
test what they wanted to check: that after dequeueing (which happens
because an execution finished), something happens to the
formerly-queued execution.

In the case of "test_queue_before", the something is, parameters
get reevaluated.
And in the case of "test_queue_nonexistent", the something is, the
workflow fails.