RD-4400 Scale rollback: keep preexisting instances around (#3593)
* RD-4400 Scale rollback: keep preexisting instances around

We can't just delete all of them and then recreate. That of course
would cascade and delete all related objects (eg: agents table).

Instead, do the usual 3-step program of "delete the ones that are
missing from old, add the ones that are missing from new, update
otherwise".

Also add an inte-test

* dockercompute: if cloudify_agent is missing, don't even bother

No need to retry this task however many times. It's either there,
or it's not. Many of these tests cause a situation in which the agent
errors out, of course.