RD-4557 inte-tests: create valid admin tokens (#3595)
* get_token_status: don't break on dashes in the token secret

We don't really need to assume that the token secret doesn't contain
dashes. Usually it won't, but why depend on it.

* auto-created tokens: make the description a bit more descriptive

execution id is very required in there. There can be many executions
for the same workflow id, we need a way to distinguish them

* inte-tests: create tokens using the same method as in the restservice

uuids arent necessarily valid (well, they are now). But still, let's
use the same method.

* flake

Co-authored-by: geokala <geokala@users.noreply.github.com>