RD-4502 bp validate inte-tests: assert on exc.error
We now show the error in just execution['error'], and the UI (and CLI)
display that. So let's assert on that in the test.

Also, since #3608 - the error isn't even in events, it's just in the
exc['error'].
Should we have left it there? Dunno. Maybe. Anyway if it's in events,
then it's printed out twice (actually, 3 times), in the CLI. Which
just looked bad. Hm.