RD-4615 Extend valid values based on type hierarchy
Extend any provided list of `valid_values` for the node types based on
the type hierarchy.  This way also descendant types will be treated as
valid.

For example for a node of type `type1` which derives from the `root`
type (so a database representation of that node will have fields:
`type=type1, type_hierarchy=[root, type1]`) and `valid_values`
constraints containing value `root` (but not `type1`), the search will
return that `type1` type because it derives from the `root` listed as a
valid value.