RD-4711 Use current deployment_id as default constraint value
When running a workflow with a parameter of data-based data type, which
requires `deployment_id` constraint, but no value was provided (in a
blueprint), use an identifier of a deployment the workflow is run on.

This is some kind of default fall-back value, which is probably also
going to be the most common one.