RD-4627 Use a long-running connection for amqp (#3643)
* workflow_executor: use a long-running connection for workflows

Sending a workflow, we used to make a new amqp connection.

That's very wasteful. Making a connection takes a long time (this is
with TLS, remember).

Instead, let's keep a long-running connection and keep reusing that.
Way lower latency for sending workflows.

* workflow_executor: keep reconnecting forever

Now that the connection is supposed to be long-running, we'll
never gonna give up reconnecting. And never let you down.

* Also do the same for hooks

Use a long-running connection for hooks too. You probably guessed this
is going to happen, because the earlier commit did include creating
a hook handler :)

* update test mocks

* remove disused send_mgmtworker_task

* also use the long-running connection for service tasks

* flake

Co-authored-by: geokala <geokala@users.noreply.github.com>