StorageManager: allow setting user&tenant
We want to be able to use the storage manager outside of requests
(eg. in tests), so we need a way to provide to it the "current" user
and tenant.

Before, we used to set request_context in tests, just for this, but
that's kinda stupid. Calls in tests aren't in the context of a request,
there's no need to pretend they are.