RD-4826 Only load the user in user_loader
Now this is the interesting one!

I'm drawing a clear divide between user_loader and authenticate here:
user_loader examines the request and loads the user, while authenticate
just requires that there is a logged-in user.

Context you might need to understand this:
- user_loader is set as a flask-login "request_loader", ie. it'll be
  called with the request, when the current_user proxy is first
  accessed
- flask-principal makes sure to access current_user on every request,
  so our user_loader is essentially always called before a request
- authenticate is a decorator used for SecuredResource

So now, user_loader is where the important stuff happens. We load the
user based on the 4 sources defined there.