RD-4826 Deduplicate user loader (#3651)
* StorageManager: allow setting user&tenant

We want to be able to use the storage manager outside of requests
(eg. in tests), so we need a way to provide to it the "current" user
and tenant.

Before, we used to set request_context in tests, just for this, but
that's kinda stupid. Calls in tests aren't in the context of a request,
there's no need to pretend they are.

* Tests: create .sm with the current user

Now actually using the previous commit. The storage-manager will have
the user baked-in, and so request-context is now not necessary for
sm calls in the tests.

* base_test: also allow passing _different_ users to sm

multi-tenant tests (mostly in cloudify-premium) also use rest-clients
with different users/tenants to do .put_deployment, so now that we
bake in the user into the SM, let's bake in the _correct_ user,
ie. the same user that is in the restclient.

This part is kinda ugly, unfortunately.

* resolver_with_catalog_support: store SM on the instance

We can only get the SM once. Also, this will allow creating the
resolver in tests with a pre-configured SM.

* filters infra: pass through SM

Instead of getting the storage manager from global somewhere deep
in the stack, pass it through. It's just nicer. Also will allow us
to pass a pre-configured SM in tests.

* reduce the dependency on global current_user

And now, actually using the stuff implemented in earlier commits:
- pass through sm where applicable
- stop using current_user where the functions can be called outside
  request context

* rest_utils: allow passing in the user as an argument

No need to always rely on current_user. Sometimes (in sm) we want to
pass in the user object explicitly. This allows calling those functions
outside of request context

* RD-4826 Only load the user in user_loader

Now this is the interesting one!

I'm drawing a clear divide between user_loader and authenticate here:
user_loader examines the request and loads the user, while authenticate
just requires that there is a logged-in user.

Context you might need to understand this:
- user_loader is set as a flask-login "request_loader", ie. it'll be
  called with the request, when the current_user proxy is first
  accessed
- flask-principal makes sure to access current_user on every request,
  so our user_loader is essentially always called before a request
- authenticate is a decorator used for SecuredResource

So now, user_loader is where the important stuff happens. We load the
user based on the 4 sources defined there.

* Use the `authenticate` function

We dropped the idea of the "authenticator" object, there's now just
a function

* CYBL-1620 Don't even bother loading the user for unauth'd requests

Those endpoints aren't going to require a user, no need to load it.
Loading it is in principle just wasteful, but also, in case of
incorrectly-written external auth, could just break the call.

* flake

* Only allow the bootstrap admin to bypass ext auth