RD-4640 Set NI status & drift when installing it (#3656)
* test_operations: add more utils

This is pretty much a copy of what we have somewhere else. Eventually
I'll want to consolidate it, but not yet. For now, let's keep per-test
implementations of those.

* Add tests for on_success operations

* _on_success_SendNodeEventTask: return on invalid params

the `pass` would've just led to a keyerror later

* make sure to lock the NI

we're going to be modifying it!

* RD-4640 Set NI status & drift when installing it

Automatically set those properties. Values TBD, but for right now,
they're this.

* elif