REST tests: select the expected user/tenant
We can't just .first without an ordering. What user would that be?
There's more than one in the db after all (eg. admin and
status_reporter).

I think this is why we've been seeing pretty-rare flakiness on the
tests, where eg. in the config tests, the updater was a different
user than we expected.