RD-4408 Instance drift/status counts (#3700)
* Test drifted/unavailable counts

That's how this is supposed to work: we set system_properties, and the
instance counts on the node and on the deployment are updated as well.

* Recompute is_ok/has_drift on update

On update, and also when creating a new NodeInstance python-side,
set those values.

Somewhat unrelatedly, I also put the nodeinstance stuff under an
explicit transaction. We do locking=True, so might as well clearly
show what's the scope of that lock.

* Replace response-only fields with actual db fields

Those values will now be stored, not just computed on demand.
Thanks to this, we'll be able to list them along with the deployments
list, and also filter on it.

* Migrations & triggers

This is the the interesting one!

I'm adding the new columns, and also triggers that keep them in sync.
Hopefully the UPDATE calls are self-explanatory.

* Default is_status_ok to false if empty

* flake

* Fix that one test

THATS IT i'm removing this assertion. After bumping it about
twenty times over the years.

* Replace the compute_* methods with update_*

Co-authored-by: Mateusz Neumann <mateumann@gmail.com>