RD-4429 REST tests: clean up .put_deployment_with_labels (#3711)
* drop BaseTest.assertEmpty

The comment is 1) untrue, 2) in the wrong place. Anyway, I can easily
just replace the thing with a `assert not`. The `is not None` I'm
gonna say is not important.

* create-dep-env: cope with empty nodes

Together with cloudify-cosmo/cloudify-common#1073 this allows
for creating a blueprint with an empty plan. Which mostly only matters
in tests, but it's nice to be able to not provide all the empty
structure every time.

* Drop usages of put_blueprint_with_labels

Replace all the .put_deployment in those tests, with just making
the deployment object directly. This is WAY faster and this commit
alone speeds up the tests by... well, a bit.

* clean up a few warnings

- node_name is deprecated in favour of node_id
- callable_obj is deprecated too, in favour of positional arg
- pytest attempts to collect TestClient even though it's not a test

* compare to empty list

* ..also compare to list in the other cases