RD-4429 deployment labels: don't update computed labels (#3728)
Some labels (currently only csys-consumer-id) are computed by the
manager, so setting them by the user doesn't make sense.

Originally the idea was to throw a 4xx error when the user attempts
to change that label (add or remove), but I think it's better to
just ignore that label.

This is because the update_labels rest call wants the user to provide
the set of the new labels (which it must, to allow deletion). So we'd
be requiring the user to always pass through the computed label
unchanged. That seems silly. Let's allow the user to pass it, or not
pass it, and we'll just keep the label as it needs to be anyway.