After install, set NI .is_status_check_ok
In the operation update, when running the install workflow, we do
set node instance's system_properties to declare that the status is OK
(and there's no drift).

However this info is also denormalized into a separate
is_status_check_ok (and has_configuration_drift) column, so we have
to update that column as well.

This will then also update unavailable_instances (and drifted_instances)
counts on the deployment (via trigger).