Rewrite reinstall_instances
This is the interesting one! Now, instead of just always reinstalling
the instances, we'll also run check_drift beforehand, and also run
the update interfaces as well.

Well, not much to say here. The implementation is pretty much new.

Relationships update is TODO for now, it'll happen in another PR.