RD-4466 update_instances: run the relationships update as well
needs cloudify-cosmo/cloudify-common#1105

This actually runs the update on relationships as well.
I don't have anything smart to say about this. It is what it is.

Anyway, the lifecycle method now examines the system_properties to
see what kind of drift do we have (own, or relationship).
Therefore, for instances that changed, but don't report any
drift (because they dont have check_drift), we add a "fake" drift
marker.