RD-4466 dep-up: add skip_drift_check (#3755)
* update_or_reinstall_instances: use install_params

rather than passing however-many-arguments to this function, pass just
the config object that we already gathered, into it.

* dep-up endpoint: shorten formatting params

a trivial refactor

* clean up the skip_reinstall type confusion

It is a bool! Not a list! It's a bool in the rest docs, the tests,
and the cli. It's not a list.

* Implement skip_drift_check

Does what it says on the tin! Skips the drift check. Duh.

* add a test

Same driver code, different blueprint