RD-5353 dep-up: skip updating uninitialized instances (#3775)
* tests: install the deployment before updating

but omit the operations run during install in the operation check.
No need to assert on install. We already know install works.

* avoid running check_drift on the whole deployment

empty set passed means that we must just run nothing.

* RD-5353 update: skip not installed instances

That's a change and is in principle backwards-incompatible, but this is
what product wants. Also I'm not much worried about back-compat for
this, because I consider updating a non-installed deployment a pretty
much invalid operation anyway.

* roll to_skip into changed_instances

remove it up front and we don't need to worry about it later.

I'm still going to leave it around in the later `must_reinstall -= to_skip`,
because maybe a skipped instance was part of a contained subgraph of
something.

* add a test

this specialcase NI state thing is pretty stupid, but we can't very
easily make it work in c-common/local.py, because all the tests ever
assume that state is None there. Pfff.