RD-5434 node-instance install: always clear drift/status
Instead of setdefault, just set it. Otherwise, if the node was
considered eg. having drift before the installation, it would still
have it. That doesn't make sense. Installation is supposed to leave
the node pristine and new.

This is mostly only visible in reinstalls (especially called by heal).
Without this, we had eg. the component heal operation, run more times
than necessary (it was still retried after a reinstall, until it
got around to running check_status again)