inte-tests: use --lightweight to skip some services (#3784)
* RD-5465 status: mark stage optional

There is a way to not install it, so it's clearly optional.

* inte-tests: use --lightweight to skip some services

* test_status_response: no need to touch prometheus

the status response is only based on a supervisord query, doesnt
need to touch prometheus

* test_status_response: stage/composer/etc are optional

they wont necessarily be there. Let's just assert on the items we
know are always there.

* drop test_status_optional_service_inactive

there's no difference between that an any other service. Let's just
not have this test.

* test_status_service_inactive: only assert on one service

why do the same thing 3 times.

* flake