restservice: allow mypy typechecking the codebase (#3787)
* Add a whole bunch of typing declarations

Those are pretty trivial, usually `Dict[str, Any]` declarations that
are not very interesting but happen all around the place.

Feel free to only skim this commit

* Handle Optional[] by adding checks

It's a type error to use Optional[T] as just T. There needs to be a
None check.

* settle on namedtuple names

mypy does care that the namedtuple is named the same, as the name
it is assigned. Funny.

* ignore typechecks in some places

Sometimes mypy is just wrong, sometimes we're using private attributes

* typecheck around conditional imports

This is the only way I've found to actually work

* fix test_post_provider_context_twice_fails

* fix _endpoint_strip_version

an actual bug!

* fix get_version_data

a type error!

* test_finalize_error_if_deployments_not_updated: run asserts

those asserts weren't run, but now they are. And now they fail.
I'm not sure what to do about them.

* storage manager: inline a method

I don't really like to do this, but this satisfies the typechecker.
Otherwise it doesn't know that the exception is of this type.

* typecheck hybrid_property

mypy only allows `@property` to have a setter of the same name as
a getter. But hybrid_property wants that too! This is a bit silly,
but until mypy fixes their shit, hybrid properties need to be declared
as properties for typechecking

* re-expose sqlalchemy models as module-level names

This is REALLY stupid. I hope mypy fixes that. Anyway,
`class X(db.Model)` seems to be just disallowed. Needs to be
`class X(Model)` instead. Stupid.

* drop unused method

* stop using deprecated attribute

exc.message has been LONG deprecated

* decide on base state for _map_execution_to_plugins_update_status

this method needs to return a str, not Optional[str].
Well, the non-covered cases are when the execution is in progress.
So let's return an "in progress" state in those cases.

* remove seemingly unused branch

well, ._modification_breadcrumbs is an AttributeError, and so
is .modification_breadcrumbs - weird.
Anyway, this module is currently unused anyway, and let's make sure to
drop all of the dep_up/handlers.py & manager.py for 7.0

* import sqlalchemy exceptions from the right place

this is where sqlalchemy declares them to be (in their type decls)

* update_plugin_imports: use a class rather than dicts

the dicts are just impossible to type, man. This is easier.

* update_plugin_imports: fixup type declarations

* update_plugin_imports: black-format

Man, I don't really feel like formatting that big dict by hand.
Lemme just run black on it. It looks fine to me.

* add a mypy.ini

* responses: make self.items non-optional

now it's just always a sequence

* import quote from requests.compat

that is the place requests declares it exports it from

* version.py: fix community version data

This needs to be a string, not bytes! Also, package 'cloudify'
doesn't exist anyway. 'cloudify-rest-service' does.