drop skipped test_list_agents_all_tenants
It was always skipped because `find_spec('cloudify_premium')` is
always None, because the name is cloudify-premium, not cloudify_premium.

Anyway, this test is just not worth it. The behaviour we're testing here
is listing with all_tenants=True, which is not specific to agents at
all, most every endpoint has it, and we test it in general in the
restservice infrastructure tests.

So, spend several minutes (this test installs agents!) on that... no.