Stop skipping several inte-tests (#3798)
* drop skipped test_list_agents_all_tenants

It was always skipped because `find_spec('cloudify_premium')` is
always None, because the name is cloudify-premium, not cloudify_premium.

Anyway, this test is just not worth it. The behaviour we're testing here
is listing with all_tenants=True, which is not specific to agents at
all, most every endpoint has it, and we test it in general in the
restservice infrastructure tests.

So, spend several minutes (this test installs agents!) on that... no.

* Unskip dep-up description tests

I think old dep-up didn't change the description, but the new one does.
So these tests just work.

* Delete test_snapshots_events

Well, it doesn't cause deadlocks anymore, I think.

But, the only thing it tests is not really anything about the snapshot
itself, but about how the snapshot _execution_ works.

All the snapshot tests check that the execution finishes. I don't want
to add one that does nothing more. It takes several minutes to run.

* flake