restservice: update flask dependencies (#3802)
* inte-tests: drop --service-management

We're never testing on systemd. It's time to stop parametrizing it.
System-tests can keep testing it, it's a way more system-related
thing anyway.

* Upgrade flask to 2+

This is seemingly the only change that's needed: itsdangerous now
needs the secret key up front.

* (temporarily) disable swagger

With these versions of dependencies (jinja2), flask-restful-swagger
will just not work. So I'm getting rid of it. RD-5683 is about
bringing it back.

* install gunicorn in the .spec

Why setup.py? lol? We only need gunicorn for deploying the restservice,
not for eg. unittests

* update restservice requirements

- flask-restful doesn't matter much, should always be the newest.
  Anyway I'm going to drop it soon.
- flask-login can be just gone, it's part of flask-security anyway
  (and soon, we'll revert back to having ONLY flask-login,
   not flask-security)
- virtualenv is not required at all
- werkzeug/itsdangerous are just requirements of flask

* regenerate reqs.txt

* flake

* RD-5683 Prerender swagger docs

Since we're temporarily taking swagger down, let's prerender the jsons
so that the UI can still hit them before we solve RD-5683

* drop SecuredResourceBannedSnapshotRestore

It's pointless. Old status reporters are long gone.

* rewrite test_rest_decorators

Don't mock flask.request, it doesn't work w/ flask 2+ apparently.
Instead, this is the intended way.

* stop using removed .reload_user