workflow update test: rewrite using an update handler (#3805)
This operation might end up updating the same node multiple times
concurrently, if it's run on the node, and at the same time, on the
relationship.
Then, we get a version conflict, and test_workflow would (rarely)
fail.
To avoid that, use the .update(on_conflict) variant of updating
node instances.

Funnily, I was long looking for this race in the wrong place: I thought
that it was due to insufficient locking in c-common's drift update
of system-properties, but no, that one is locked just fine. It's
the _runtime_ properties, in the test operations, that were at fault!