fixup the deployment overrides test (#3810)
* rest tests: add missing permissions

we are explicitly checking those permissions, let's declare them

* deployments override: don't forget display name!

* dep overrides test: correct scaling_groups & workflows syntax

* visibility is not just a random templated string

* dep overrides test: getattr

* display name: throw early

as test_deployment_display_name_with_control_chars_fails noted,
we used to only throw this error in the later set-attributes call,
but we _can_ in this case throw it up front. So let's do it.

The main deployments PUT method is now even uglier than it was. Wants
splitting to several methods probably.