Deployment overrides (#3808)
* No need to conceal init behind a nop

* Update created_by without traceback

* Allow setting executions

* Add tests for overriding deployment params

* Remove outdated test

The architecture that made this problem possible was changed some time ago,
so let's not keep a test that tries to have a little race around.

* Move test for dep create overrides

* fixup the deployment overrides test (#3810)

* rest tests: add missing permissions

we are explicitly checking those permissions, let's declare them

* deployments override: don't forget display name!

* dep overrides test: correct scaling_groups & workflows syntax

* visibility is not just a random templated string

* dep overrides test: getattr

* display name: throw early

as test_deployment_display_name_with_control_chars_fails noted,
we used to only throw this error in the later set-attributes call,
but we _can_ in this case throw it up front. So let's do it.

The main deployments PUT method is now even uglier than it was. Wants
splitting to several methods probably.

Co-authored-by: Lukasz Maksymczuk <tehasdf@users.noreply.github.com>