Rewrite rest/test_plugins.py
Well, it's a lot of changes, sorry about that, but I'm not gonna split
this per change type after all. So this is going to be a kinda bad
commit. Anyway...

Rewrite all the tests in test_plugins.py so that they don't use
`self.upload_plugin` (which builds a wagon), but rather,
a few "mocking" techniques instead:

1. Where we test something related to already-stored plugins, eg.
   listing, just create the entries in the db, via `models.Plugin()`
2. When testing something related to upload itself, instead of creating
   a wagon, just create a tempfile. It doesn't really need to keep
   a wagon structure all that much, but where we do care about files,
   create some dummy files in it too.
3. Mock out the `wagon.show` call. Wagon is not part of the unit under
   test here. This way we can pretend the plugin contains whatever
   package name we want!