RD-5880 rest tests: rewrite plugin tests to not make wagons (#3817)
* Rewrite rest/test_plugins.py

Well, it's a lot of changes, sorry about that, but I'm not gonna split
this per change type after all. So this is going to be a kinda bad
commit. Anyway...

Rewrite all the tests in test_plugins.py so that they don't use
`self.upload_plugin` (which builds a wagon), but rather,
a few "mocking" techniques instead:

1. Where we test something related to already-stored plugins, eg.
   listing, just create the entries in the db, via `models.Plugin()`
2. When testing something related to upload itself, instead of creating
   a wagon, just create a tempfile. It doesn't really need to keep
   a wagon structure all that much, but where we do care about files,
   create some dummy files in it too.
3. Mock out the `wagon.show` call. Wagon is not part of the unit under
   test here. This way we can pretend the plugin contains whatever
   package name we want!

* upload_manager: a bit more lenient on the types

those .get are pointless if we fail on None anyway, so let's
default to something that'll make it not fail.

* snapshot tests: dont create wagons

those tests just check uploading, and they want "a big file", not
necessarily a wagon (which is not a snapshot anyway). So just make
a file.

* test_deployments: assume plugins are already uploaded

No need to test plugin upload in these

* rewrite other tests the same way!

* remove now-unused methods

* drop a bunch of now-disused yaml files