reformat getting request args
the `.get(..., '{}')` is a bit silly: we're providing a literal, that
we then want to still parse. Heh. I prefer not doing that.

Also, dict()-ing request.args is equivalent to `.to_dict(flat=True)`,
which flattens the structure. We only care about the difference in
labels, because that's a list.
Before (flat=True):
`?labels=a&labels=b` -> 'a'  (only one entry is left)
After (flat=False):
`?labels=a&labels=b` -> ['a', 'b']  (a list instead)