Multipart blueprint upload: PR review changes (#3831)
* Skip the 'Transfer-Encoding' part

Doesn't seem to be the case ever.
Old-client: request.data is set, no Transfer-Encoding;
New-client: request.files is set, no Transfer-Encoding
Browser: well, still request.data, but that'll change. No T-E.
Passing in a url: no req.data, no T-E, no req.files.

Anyway, in principle, transfer-encoding can be set indeed, but then,
one of the other things (req.data or req.files) is going to be
set as well. So no need to check this specifically.

* reformat getting request args

the `.get(..., '{}')` is a bit silly: we're providing a literal, that
we then want to still parse. Heh. I prefer not doing that.

Also, dict()-ing request.args is equivalent to `.to_dict(flat=True)`,
which flattens the structure. We only care about the difference in
labels, because that's a list.
Before (flat=True):
`?labels=a&labels=b` -> 'a'  (only one entry is left)
After (flat=False):
`?labels=a&labels=b` -> ['a', 'b']  (a list instead)

* revert the generate_progress_func change

We can't (and are not) using this func with multipart upload progress
anyway, because it can't really work with it: the chunk/buffer size
assertion would fail, because in tests, it's all read in a single
call.

So since we're not using it, no need to make this assertion weaker (yet)