RD-5692 Temporarily skip windows & altarch agents (#3835)
For CI to work in cloudify-agent, those packages need to be built
at all. But currently the jobs that make them, don't work. To make
those jobs work, we need to be able to commit into cloudify-agent,
and to do that, CI must pass there.

This is to be reverted after those jobs pass.