restservice: factor out the entrypoint into a separate script
the point is, we can pass in the required settings as envvars, and
render the config file at startup. Obviously, we don't want to keep
the "copy config file at build time" idea, and volumes are not as
nice to use as just envvars.

Also, have some default values for the envvars. Maybe only for now.