rest-service: rewrite some tests to not use .put_deployment (#3854)
I'd like to stop using .put_deployment and .put_blueprint in these
restservice tests. Those tests are supposed to be testing the
restservice, and parsing blueprints + creating deployments, is out
of scope for that.

Instead of parsing blueprints & creating deployments, we'll create
the things we need directly: a deployment entry in the DB, and where
needed, Node and NodeInstance entries.

This means that unfortunately many tests are going to be more verbose,
but on the upside, they're going to be:
 - self-contained (no separate .yaml)
 - MUCH faster
 - more direct (just create things that you need ONLY)

So this commit changes some tests, dropping blueprint upload in them
in favour of creating the resources directly. It's essentially
the same change multiple times. It's not all of it, we'll have to
do this many more times in other places, but I wanted to already
start on it.