Rewrite configure_manager without create_default_user_tenant_and_roles
I think it's better if the setup/configuration stuff is right here,
not in some "storage utils". Creating the admin isn't a storage util,
is it :P
This way we have better control of what is created and what isn't,
and so we can create the tenant, the user, and the assoc separately.
And the rabbitmq stuff too.