RD-6073 Factor out creating the admin user to configure_manager (#3865)
* Rewrite configure_manager without create_default_user_tenant_and_roles

I think it's better if the setup/configuration stuff is right here,
not in some "storage utils". Creating the admin isn't a storage util,
is it :P
This way we have better control of what is created and what isn't,
and so we can create the tenant, the user, and the assoc separately.
And the rabbitmq stuff too.

* Add some tests

* drop create_default_user_tenant_and_roles from storage_utils

We can just get rid of it now, then. It's still used in tests,
but we'll work on that next.

* Rewrite base_test to not use the removed util

It's also a bit more sane, because now we only create the
user/tenant right there, instead of creating it in one method, and then
also retrieving it immediately after

* drop now-unused stuff from security_utils

I also dropped the status_reporter users cruft in the previous
commit... well, those have been gone since like Cloudify 5.0

* inte-tests: stop using create_default_user_tenant_and_roles

The script that resets the db between tests can be now made... simpler.
We don't really _need_ to re-create the rabbitmq stuff either. It'll
be faster this way.

* only set rabbitmq_password in the one test that needs it

no need to carry that very big string everywhere with us. Also,
assert on the actual value too, not just that it's not none